Use BIGINT for node_id

+++ ptproc-pgsql.sql
@@ -1,4 +1,4 @@
-CREATE TABLE pt_nodes (node_id INT PRIMARY KEY,bus_ref TEXT,trolleybus_ref TEXT,tram_ref TEXT,share_taxi_ref TEXT);
+CREATE TABLE pt_nodes (node_id BIGINT PRIMARY KEY,bus_ref TEXT,trolleybus_ref TEXT,tram_ref TEXT,share_taxi_ref TEXT);
 CREATE TABLE pt_ways (way_id INT PRIMARY KEY,bus_ref TEXT,trolleybus_ref TEXT,tram_ref TEXT,share_taxi_ref TEXT);
 CREATE TABLE pt_routes (
   master_id INT,
